requirements.txt
Django~=3.2.16
djangorestframework~=3.14.0
PyMySQL~=1.0.2
djoser
django-cors-headers
django-filter